+++ Sources/Runner/main.swift
@@ -1,9 +1,12 @@
 import Foundation
 
 let cliLength = ProcessInfo.processInfo.arguments.count
-print(ProcessInfo.processInfo.arguments)
-if cliLength > 1 && "edit" == CommandLine.arguments[1] {
-    try editDanger()
-} else {
-    runDanger()
+do {
+    if cliLength > 1 && "edit" == CommandLine.arguments[1] {
+        try editDanger()
+    } else {
+        try runDanger()
+    }
+} catch {
+    exit(1)
 }